requirements.txt
black==22.10.0
Django==4.0.0
django-cleanup==6.0.0
django-storages==1.11.1
django-allauth==0.45.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery-results==2.4.0
gunicorn==20.1.0
mysqlclient==2.1.1
django-redis==5.2.0
dj-rest-auth==2.2.5
django-environ==0.9.0
django-cors-headers==3.13.0
oauth2client==4.1.3
drf-spectacular==0.24.2
django-filter==22.1
